changing isSystem to canSupport and adding UI for initial Lab result
