+++ src/services/fhir/types.ts
@@ -1,6 +1,6 @@
 interface  ObservationValidator {
   /* this checks if observation system could be applied */
-  isSystem(entry: BundleEntry): boolean 
+  canSupport(entry: BundleEntry): boolean 
   validate( entry: BundleEntry ): boolean
 }
